cut off at current project
