+++ packages/checkout/sdk/package.json
@@ -53,7 +53,7 @@
   "scripts": {
     "build": "rollup --config rollup.config.js",
     "build:dev": "CHECKOUT_DEV_MODE=true yarn build",
-    "dev": "CHECKOUT_DEV_MODE=true rollup --config rollup.config.js -w",
+    "dev": "swc src -d dist --strip-leading-paths --ignore '**/*.test.*'",
     "docs": "typedoc --plugin typedoc-plugin-markdown --skipErrorChecking --disableSources --out docs src/index.ts",
     "lint": "eslint ./src --ext .ts,.jsx,.tsx --max-warnings=0",
     "lint:fix": "eslint ./src --ext .ts,.jsx,.tsx --max-warnings=0 --fix",
@@ -67,5 +67,5 @@
   },
   "source": "src/index.ts",
   "type": "module",
-  "types": "dist/index.d.ts"
+  "types": "src"
 }
