+++ packages/internal/generated-clients/package.json
@@ -23,12 +23,12 @@
   "repository": "immutable/ts-immutable-sdk.git",
   "scripts": {
     "build": "NODE_ENV=production rollup --config rollup.config.js",
-    "dev": "rollup --config rollup.config.js -w",
+    "dev": "swc src -d dist --strip-leading-paths --ignore '**/*.test.*'",
     "prepare": "wsrun -r build",
     "test": "jest",
     "typecheck": "tsc --noEmit --jsx preserve",
     "view-generators": "openapi-generator-cli author template -g typescript-axios -o src/templates"
   },
   "type": "module",
-  "types": "dist/index.d.ts"
+  "types": "src"
 }
