+++ packages/internal/guardian/package.json
@@ -38,8 +38,8 @@
   "repository": "immutable/ts-immutable-sdk.git",
   "scripts": {
     "build": "NODE_ENV=production rollup --config rollup.config.js",
-    "dev": "rollup --config rollup.config.js -w"
+    "dev": "swc src -d dist --strip-leading-paths --ignore '**/*.test.*'"
   },
   "type": "module",
-  "types": "dist/index.d.ts"
+  "types": "src"
 }
