+++ packages/webhook/sdk/package.json
@@ -35,7 +35,7 @@
   "repository": "immutable/ts-immutable-sdk.git",
   "scripts": {
     "build": "NODE_ENV=production rollup --config rollup.config.js",
-    "dev": "rollup --config rollup.config.js -w",
+    "dev": "swc src -d dist --strip-leading-paths --ignore '**/*.test.*'",
     "generate-types": "typechain --target=ethers-v5 --out-dir=src/typechain/types 'abi/*.json'",
     "lint": "eslint ./src --ext .ts,.jsx,.tsx --max-warnings=0",
     "test": "jest --passWithNoTests",
@@ -43,5 +43,5 @@
     "typecheck": "tsc --noEmit --jsx preserve"
   },
   "type": "module",
-  "types": "dist/index.d.ts"
+  "types": "src"
 }
